支持构建arm64架构的成果物

+++ build.sh
@@ -2,6 +2,18 @@
 
 set -e
 
+ARCH=$1
+
+if [ -z "$ARCH" ]; then
+    echo "错误：未指定ARCH参数。"
+    exit 1
+fi
+
+if [ "$ARCH" != "amd64" ] && [ "$ARCH" != "arm64" ]; then
+    echo "错误：ARCH参数必须是amd64或arm64。"
+    exit 1
+fi
+
 rm -rf ./bin/ || true
 mkdir ./bin
 
@@ -9,7 +21,7 @@ BUILD_NAME=my-ebpf
 VERSION=$(cat version.txt)
 IMAGE_NAME=${BUILD_NAME}:${VERSION}
 
-docker buildx build -t ${IMAGE_NAME} --platform=linux/amd64 -o type=docker .
+docker buildx build -t ${IMAGE_NAME} --platform=linux/${ARCH} -o type=docker .
 
 CID=$(docker create ${IMAGE_NAME})
 docker cp ${CID}:/usr/bin/my-ebpf ./bin/
